welcome/index: JSをStimulusのコントローラに移動する
