+++ app/assets/javascripts/application.js
@@ -31,7 +31,7 @@ $(function () {
   'use strict';
 
   var $body = $('body');
-  var controller = $body.data('controller').replace(/\//, '_');
+  var controller = $body.data('rails-controller').replace(/\//, '_');
   var action = $body.data('action');
   var activeController = logArchiver[controller];
