+++ app/views/layouts/application.html.erb
@@ -17,7 +17,7 @@
       <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
     <![endif]-->
   </head>
-  <body data-controller="<%= controller.controller_path %>" data-action="<%= controller.action_name %>" data-view="<%= yield(:view) %>">
+  <body data-rails-controller="<%= controller.controller_path %>" data-action="<%= controller.action_name %>" data-view="<%= yield(:view) %>">
     <%= render('shared/navbar') %>
 <%= yield %>
   </body>
