+++ app/views/welcome/index.html.erb
@@ -1,5 +1,5 @@
 <% provide(:view, 'welcome/index') %>
-<div class="container main-container">
+<div class="container main-container" data-controller="welcome--index">
   <div class="row">
     <div class="col-xs-12 col-md-8 col-md-offset-2">
       <div class="main-panel">
